refactor(inbox): split store.ts into store/{queries,confirm,plans,scan}.ts + extract inbox-list-model.ts
C-14: inbox store.ts (980L) split into 4 focused modules re-exported from
store/index.ts — import sites unchanged.

C-18: InboxList.tsx (851L) pure row-model layer (sort, filter, classification
labels, visual-row tree flatten) extracted to inbox-list-model.ts. InboxList
keeps only React rendering; re-exports types/constants so external imports
are preserved.

Behavior-preserving, typecheck green.
